+++ examples/Model-method-efficacyFunction.R
@@ -4,11 +4,11 @@ my_data <- DataDual(
 )
 
 my_model <- Effloglog(
-  Eff = c(1.223, 2.513),
-  Effdose = c(25, 300),
+  eff = c(1.223, 2.513),
+  eff_dose = c(25, 300),
   nu = c(a = 1, b = 0.025),
   data = my_data,
-  c = 2
+  const = 2
 )
 
 eff_fun <- efficacyFunction(my_model, theta1 = -4.8, theta2 = 3.7)