+++ examples/Model-method-gain.R
@@ -1,3 +1,5 @@
+# nolint start
+
 ##Obtain the gain value for a given dose, a pseudo DLE model, a DLE sample,
 ## a pseudo efficacy model and an efficacy sample
 ##The DLE model must be from 'ModelTox' class (DLEmodel slot)
@@ -8,8 +10,10 @@ DLEsamples <- mcmc(data, DLEmodel, McmcOptions(burnin=100,step=2,samples=200))
 
 ##The efficacy model must be from 'ModelEff' class (Effmodel slot)
 ## The DLE and efficacy samples must be from 'Samples' class (DLEsamples and Effsamples slot)
-Effmodel<-Effloglog(Eff=c(1.223,2.513),Effdose=c(25,300),nu=c(a=1,b=0.025),data=data,c=0)
+Effmodel<-Effloglog(eff=c(1.223,2.513),eff_dose=c(25,300),nu=c(a=1,b=0.025),data=data)
 Effsamples <- mcmc(data, Effmodel, McmcOptions(burnin=100,step=2,samples=200))
 
 ## Given a dose level 75,
 gain(dose=75,DLEmodel=DLEmodel,DLEsamples=DLEsamples,Effmodel=Effmodel,Effsamples=Effsamples)
+
+# nolint end