+++ NEWS.md
@@ -1,4 +1,6 @@
 # Version 1.0.9000.9012
+* Renamed `Eff` argument to `eff` and `Effdose` to `eff_dose` in `Effloglog` class
+  constructor.
 * Renamed `ExpEff` methods for pseudo-data efficacy models to `efficacy`.
 * Moved `dose` and `ExpEff` functions from model class slots to model class
   methods for pseudo-data efficacy models.